tests, network: Move `vmi_servers` to a new `network/vmnetserver` pkg
Move the helpers closer to its common users and maintainers.
