+++ tests/libnet/vmnetserver/BUILD.bazel
@@ -1,0 +1,15 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["servers.go"],
+    importpath = "kubevirt.io/kubevirt/tests/libnet/vmnetserver",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//tests/console:go_default_library",
+        "//vendor/github.com/google/goexpect:go_default_library",
+        "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
+    ],
+)
