+++ tests/libnet/vmnetserver/servers.go
@@ -1,12 +1,32 @@
-package tests
+/*
+ * This file is part of the KubeVirt project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Copyright 2020 Red Hat, Inc.
+ *
+ */
+
+package vmnetserver
 
 import (
 	"fmt"
 	"strings"
 	"time"
 
-	expect "github.com/google/goexpect"
 	. "github.com/onsi/gomega"
+
+	expect "github.com/google/goexpect"
 	k8sv1 "k8s.io/api/core/v1"
 
 	v1 "kubevirt.io/api/core/v1"
@@ -64,7 +84,7 @@ EOL`, inetSuffix, port)
 	Expect(console.ExpectBatch(vmi, []expect.Batcher{
 		&expect.BSnd{S: fmt.Sprintf("%s\n", serverCommand)},
 		&expect.BExp{R: console.PromptExpression},
-		&expect.BSnd{S: EchoLastReturnValue},
+		&expect.BSnd{S: "echo $?\n"},
 		&expect.BExp{R: console.ShellSuccess},
 	}, 60*time.Second)).To(Succeed())
