+++ tests/migration/BUILD.bazel
@@ -47,6 +47,7 @@ go_library(
         "//tests/libnet:go_default_library",
         "//tests/libnet/job:go_default_library",
         "//tests/libnet/service:go_default_library",
+        "//tests/libnet/vmnetserver:go_default_library",
         "//tests/libnode:go_default_library",
         "//tests/libpod:go_default_library",
         "//tests/libstorage:go_default_library",
