+++ tests/network/BUILD.bazel
@@ -68,6 +68,7 @@ go_library(
         "//tests/libnet/dns:go_default_library",
         "//tests/libnet/job:go_default_library",
         "//tests/libnet/service:go_default_library",
+        "//tests/libnet/vmnetserver:go_default_library",
         "//tests/libnode:go_default_library",
         "//tests/libpod:go_default_library",
         "//tests/libvmi:go_default_library",
