+++ tests/network/vmi_istio.go
@@ -54,6 +54,7 @@ import (
 	"kubevirt.io/kubevirt/tests/libnet/cloudinit"
 	"kubevirt.io/kubevirt/tests/libnet/job"
 	netservice "kubevirt.io/kubevirt/tests/libnet/service"
+	"kubevirt.io/kubevirt/tests/libnet/vmnetserver"
 	"kubevirt.io/kubevirt/tests/libvmi"
 	"kubevirt.io/kubevirt/tests/libwait"
 	"kubevirt.io/kubevirt/tests/testsuite"
@@ -118,7 +119,7 @@ var istioTests = func(vmType VmType) {
 	Context("Virtual Machine with istio supported interface", func() {
 		createJobCheckingVMIReachability := func(serverVMI *v1.VirtualMachineInstance, targetPort int) (*batchv1.Job, error) {
 			By("Starting HTTP Server")
-			tests.StartPythonHttpServer(vmi, targetPort)
+			vmnetserver.StartPythonHttpServer(vmi, targetPort)
 
 			By("Getting back the VMI IP")
 			vmi, err = virtClient.VirtualMachineInstance(vmi.Namespace).Get(context.Background(), vmi.Name, &metav1.GetOptions{})
@@ -361,7 +362,7 @@ var istioTests = func(vmType VmType) {
 
 				By("Starting HTTP Server")
 				Expect(console.LoginToAlpine(serverVMI)).To(Succeed())
-				tests.StartPythonHttpServer(serverVMI, vmiServerTestPort)
+				vmnetserver.StartPythonHttpServer(serverVMI, vmiServerTestPort)
 
 				By("Creating Istio VirtualService")
 				virtualServicesRes := schema.GroupVersionResource{Group: networkingIstioIO, Version: istioApiVersion, Resource: "virtualservices"}
